chore: move VerfiedUserData to types
